Use client certificates when replicating
This patch enables the bigcouch replicator to have the same functionality
as couchdb when replicating. When connecting to a remote couch over
SSL, the replicator will look for the cert_file, key_file params in the
replicator section of the config. If they exist, the replicator will
present the credentials to the other side